requirements.txt
--extra-index-url https://test.pypi.org/simple/
flask~=1.1.2
Flask-Session
requests~=2.25.0
chess~=1.3.2
cuwais-common~=1.0.3-dev2103191
waitress~=1.4.4
google-auth~=1.24.0
cachecontrol~=0.12.6
SQLAlchemy~=1.4.1
redis~=3.5.3
sh